requirements.txt
# python 3.12.11 (doesn't work with chatterbox)
# python 3.11.13

elevenlabs
ffmpeg-python # brew install ffmpeg
hume
openai
pedalboard
pip
pydub
python-dotenv
streamlit-js-eval
soundfile
kokoro
huggingface-hub
matplotlib
streamlit
streamlit-extras
accelerate
coqui-tts # brew install xz
torchcodec
chatterbox-tts # needs local huggingface-cli token

# PARLER
# install parler-tts separately since it conflicts with streamlit-extras

# PIPER
# brew install cmake ninja pkg-config nspeak-ng
# git clone https://github.com/OHF-Voice/piper1-gpl
# ... piper venv
# pip install --upgrade pip setuptools wheel scikit-build-core cmake
# pip install -e .
# pip install build
# python -m build --wheel
# ... diatribe venv
# pip install ../piper1-gpl/dist/piper-*.whl

# End? No, the journey does not end here. Death is just another path, one that we all must take. The grey rain curtain of this world rolls back and all turns to silver glass. And then you see it.